First proper challenge in a while
By calling “answer 1000” all n in 1,000<=n<10,000 are checked. The
checking process involves generating a list of the permutations of all
of the digits, then removing all non-primes from that list. Then take
all permutations of length 3 of that list, then check each for an
arithmetic sequence. This will print a list of candidates.

Selecting the first candidate that isn’t a permutation of 1487 (which
is 2699) we can then call “find_sequences 2699”. Then just concat the
values in the first non-empty list. This makes the answer 296962999629.